Physics_Informed_UNet/requirements.txt
tensorflow==2.12.0
numpy==1.23.4
matplotlib==3.7.1
scipy==1.10.1
tqdm==4.65.0
pillow==9.5.0
scikit-image==0.20.0
pandas==2.0.1
pyyaml==6.0
